requirements.txt
# PaddlePaddle (飞桨社区环境已安装)
paddlepaddle>=3.0.0
paddlenlp>=2.6.0

# 爬虫相关
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.4.0
joblib>=1.3.0

# 工具
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
